test(settings): align frontend CI with compact settings flows
Problem: frontend CI cannot load CSS imports in five settings suites and retains outdated header and plugin installer assumptions.

Root cause: new settings shell dependencies were not registered with the Node CSS hook; the plugin test clicked the collapsed install entry instead of the install action.

Fix: register the affected suites, assert accessible shell names without duplicate headings, and exercise installer disclosure before installation. Extract plugin actions and shared DOM helpers to keep test files within their existing size limits.

Verification: all seven affected suites pass with the CI loader; repolint is clean and the full frontend regression run is in progress.
