+++ desktop/frontend/scripts/run-tests.mjs
@@ -77,7 +77,7 @@ for (const [name, owner] of OWNED_ELSEWHERE) {
 
 // Suites that statically import CSS (e.g. HeartbeatPanel's heartbeat.css) need
 // the css-stub loader hook so tsx resolves the import under node.
-const CSS_STUB_SUITES = new Set(["settings-provider-normalization.test.ts", "provider-image-input.test.tsx", "heartbeat-editor.test.tsx", "heartbeat-next-run.test.ts", "settings-page-navigation.test.tsx", "automation-management.test.tsx", "trash-management.test.tsx", "capabilities-panel-actions.test.ts", "provider-access-card.test.tsx", "provider-editor-model-picker.test.tsx", "provider-name-readonly.test.tsx", "settings-refresh-snapshot.test.tsx", "shell-support-install.test.tsx", "shortcuts-recorder-focus.test.tsx"]);
+const CSS_STUB_SUITES = new Set(["plugin-settings-actions.test.tsx", "connection-model-picker.test.tsx", "provider-display-name.test.tsx", "provider-editor-compact.test.tsx", "provider-model-discovery-selection.test.tsx", "provider-onboarding.test.tsx", "settings-provider-normalization.test.ts", "provider-image-input.test.tsx", "heartbeat-editor.test.tsx", "heartbeat-next-run.test.ts", "settings-page-navigation.test.tsx", "automation-management.test.tsx", "trash-management.test.tsx", "capabilities-panel-actions.test.ts", "provider-access-card.test.tsx", "provider-editor-model-picker.test.tsx", "provider-name-readonly.test.tsx", "settings-refresh-snapshot.test.tsx", "shell-support-install.test.tsx", "shortcuts-recorder-focus.test.tsx"]);
 
 const suites = files.filter((name) => !OWNED_ELSEWHERE.has(name));
 console.log(`run-tests: ${suites.length} discovered suites (${OWNED_ELSEWHERE.size} owned by dedicated scripts)`);
