+++ desktop/frontend/src/__tests__/capabilities-panel-actions.test.ts
@@ -1,15 +1,15 @@
-import { JSDOM } from "jsdom";
+import { findButton, flush, installDom, setInputValue, waitFor } from "./capabilities-test-helpers";
 import { readFileSync } from "node:fs";
 import React from "react";
 import { act } from "react";
 import { createRoot } from "react-dom/client";
-import { MCPServersSettingsPage, PluginsSettingsPage, failureKind, mcpServerDraftJSON, parseMCPQuickDefinition, parseMCPServerJSON, summarizeServerError, withExplicitMCPClears } from "../components/CapabilitiesPanel";
+import { MCPServersSettingsPage, failureKind, mcpServerDraftJSON, parseMCPQuickDefinition, parseMCPServerJSON, summarizeServerError, withExplicitMCPClears } from "../components/CapabilitiesPanel";
 import { slashCommandGroup, slashCommandKindTag, sortSlashCommandsForMenu } from "../components/SlashMenu";
 import { selectToolsOnFirstCustomUse } from "../components/SubagentsPanel";
 import type { AppBindings } from "../lib/bridge";
 import { LocaleProvider, t } from "../lib/i18n";
 import { mcpServerLifecycleActions, mcpServerRetryableFromAvailableList } from "../lib/mcpServerLifecycle";
-import type { MCPServerInput, Meta, PluginInstallOptions, PluginView, ServerView, TabMeta } from "../lib/types";
+import type { MCPServerInput, Meta, ServerView, TabMeta } from "../lib/types";
 
 function ok(value: unknown, message: string) {
   if (!value) throw new Error(message);
@@ -292,69 +292,6 @@ ok(!mcpServerRetryableFromAvailableList(server("connected")), "connected server 
 ok(!mcpServerRetryableFromAvailableList({ ...server("disabled"), startIntent: "off" }), "disabled server should be excluded from available-list retry all");
 ok(!mcpServerRetryableFromAvailableList({ ...server("failed"), runtimeState: "issue" }), "failed server is handled by the failure banner retry all");
 
-function flush(): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, 0));
-}
-
-async function waitFor(label: string, predicate: () => boolean) {
-  for (let attempt = 0; attempt < 20; attempt += 1) {
-    await act(async () => {
-      await flush();
-    });
-    if (predicate()) return;
-  }
-  throw new Error(`timed out waiting for ${label}`);
-}
-
-function installDom() {
-  const dom = new JSDOM("<!doctype html><html><body><div id=\"root\"></div></body></html>", {
-    pretendToBeVisual: true,
-    url: "http://localhost/",
-  });
-  (globalThis as typeof globalThis & { IS_REACT_ACT_ENVIRONMENT: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
-  globalThis.window = dom.window as unknown as Window & typeof globalThis;
-  globalThis.document = dom.window.document;
-  Object.defineProperty(globalThis, "navigator", { configurable: true, value: dom.window.navigator });
-  globalThis.Node = dom.window.Node;
-  globalThis.HTMLElement = dom.window.HTMLElement;
-  globalThis.HTMLButtonElement = dom.window.HTMLButtonElement;
-  globalThis.HTMLInputElement = dom.window.HTMLInputElement;
-  globalThis.Event = dom.window.Event;
-  globalThis.KeyboardEvent = dom.window.KeyboardEvent;
-  globalThis.MouseEvent = dom.window.MouseEvent;
-  globalThis.localStorage = dom.window.localStorage;
-  globalThis.requestAnimationFrame = dom.window.requestAnimationFrame.bind(dom.window);
-  globalThis.cancelAnimationFrame = dom.window.cancelAnimationFrame.bind(dom.window);
-  Object.defineProperty(window, "matchMedia", {
-    configurable: true,
-    value: () => ({
-      matches: true,
-      media: "(prefers-reduced-motion: reduce)",
-      onchange: null,
-      addEventListener() {},
-      removeEventListener() {},
-      addListener() {},
-      removeListener() {},
-      dispatchEvent: () => false,
-    }),
-  });
-  return dom;
-}
-
-function findButton(label: string): HTMLButtonElement | undefined {
-  return Array.from(document.querySelectorAll("button")).find((button) => button.textContent?.trim() === label) as HTMLButtonElement | undefined;
-}
-
-function setInputValue(input: HTMLInputElement, value: string) {
-  const win = input.ownerDocument.defaultView;
-  const previous = input.value;
-  const setter = Object.getOwnPropertyDescriptor((win?.HTMLInputElement ?? HTMLInputElement).prototype, "value")?.set;
-  setter?.call(input, value);
-  (input as HTMLInputElement & { _valueTracker?: { setValue: (next: string) => void } })._valueTracker?.setValue(previous);
-  const eventCtor = win?.Event ?? Event;
-  input.dispatchEvent(new eventCtor("input", { bubbles: true }));
-  input.dispatchEvent(new eventCtor("change", { bubbles: true }));
-}
 
 ok(
   slashCommandKindTag({ name: "pwf:plan", description: "Plugin planning prompt.", kind: "custom", plugin: "pwf" }, t) === "plugin · pwf",
@@ -853,252 +790,6 @@ console.log("capabilities panel MCP actions");
   });
   ok(Boolean(document.querySelector(".cap-mcp-field--name input")) && Boolean(findButton("Advanced options")), "manual setup restores name, transport, and advanced configuration without leaving the install page");
   ok(!addedInput, "opening the quick installer does not mutate MCP state");
-
-  await act(async () => {
-    root.unmount();
-  });
-  dom.window.close();
-}
-
-console.log("capabilities panel plugin actions");
-
-{
-  const dom = installDom();
-  const rootEl = document.getElementById("root");
-  if (!rootEl) throw new Error("missing root");
-  const root = createRoot(rootEl);
-  const meta: Meta = { label: "test", ready: true, eventChannel: "plugin-channel", cwd: "/tmp/reasonix-test", workspaceRoot: "/tmp/reasonix-test" };
-  const tabs: TabMeta[] = [{
-    id: "tab-plugin",
-    scope: "project",
-    workspaceRoot: "/tmp/reasonix-test",
-    workspaceName: "reasonix-test",
-    topicId: "topic-plugin",
-    topicTitle: "Plugins",
-    label: "Plugins",
-    ready: true,
-    running: false,
-    mode: "normal",
-    toolApprovalMode: "auto",
-    active: true,
-    cwd: "/tmp/reasonix-test",
-  }];
-  let planCalls = 0;
-  let installCalls = 0;
-  let toggleCalls = 0;
-  let updateCalls = 0;
-  let doctorCalls = 0;
-  let removeCalls = 0;
-  let pickFolderCalls = 0;
-  const plannedSources: string[] = [];
-  const installedSources: string[] = [];
-  let plugins: PluginView[] = [{
-    name: "superpowers",
-    version: "0.1.0",
-    description: "Shared agent skills and hooks.",
-    source: "git:github.com/obra/superpowers",
-    root: "~/.reasonix/plugins/superpowers",
-    manifestKind: "reasonix",
-    enabled: true,
-    skills: 2,
-    hooks: 1,
-    mcpServers: 0,
-  }];
-  window.go = {
-    main: {
-      App: {
-        Meta: async () => meta,
-        ListTabs: async () => tabs,
-        Plugins: async () => plugins.map((plugin) => ({ ...plugin, warnings: [...(plugin.warnings ?? [])] })),
-        PlanPluginInstall: async (source: string, options: PluginInstallOptions) => {
-          planCalls += 1;
-          plannedSources.push(source);
-          ok(options.dryRun === true, "plugin preview asks for dry-run planning");
-          return JSON.stringify({
-            ok: true,
-            status: "planned",
-            name: "superpowers",
-            actions: [{
-              kind: "plugin", action: "install_plugin_package", name: "superpowers", source, status: "planned",
-              compatibility: "partial", mappedCapabilities: ["skills", "agents"],
-              skippedCapabilities: [{ capability: "hook", path: "hooks/hooks.json", reason: "unsupported event" }],
-            }],
-          });
-        },
-        InstallPlugin: async (source: string, _options: PluginInstallOptions) => {
-          installCalls += 1;
-          installedSources.push(source);
-          const next: PluginView = {
-            name: "superpowers",
-            version: "0.1.1",
-            description: "Shared agent skills and hooks.",
-            source,
-            root: "~/.reasonix/plugins/superpowers",
-            manifestKind: "reasonix",
-            enabled: true,
-            skills: 3,
-            commands: 2,
-            agents: 1,
-            hooks: 1,
-            mcpServers: 1,
-            compatibility: "full",
-            mappedCapabilities: ["skills", "agents", "hooks", "mcp"],
-            skillDetails: [{ name: "plan", description: "Plan work before implementation.", invocation: "/superpowers:plan", runAs: "inline" }],
-            agentDetails: [{ name: "reviewer", description: "Review changes.", invocation: "/superpowers:reviewer", model: "sonnet" }],
-            commandDetails: [{
-              name: "plan",
-              description: "Plugin planning prompt.",
-              invocation: "/superpowers:plan",
-            }, {
-              name: "blocked",
-              description: "Occupied canonical command.",
-              invocation: "/superpowers:blocked",
-              shadowed: true,
-            }],
-            hookDetails: [{ event: "SessionStart", contextFile: "CLAUDE.md", description: "Load startup context." }],
-            mcpServerDetails: [{ name: "context", displayName: "Context Search", transport: "stdio", command: "node server.js", autoStart: false }],
-          };
-          plugins = plugins.filter((plugin) => plugin.name !== next.name).concat(next);
-          return JSON.stringify({ ok: true, status: "done", actions: [{ action: "install_plugin_package", name: next.name, status: "done" }] });
-        },
-        SetPluginEnabled: async (name: string, enabled: boolean) => {
-          toggleCalls += 1;
-          plugins = plugins.map((plugin) => plugin.name === name ? { ...plugin, enabled } : plugin);
-        },
-        UpdatePlugin: async (name: string) => {
-          updateCalls += 1;
-          plugins = plugins.map((plugin) => plugin.name === name ? { ...plugin, version: "0.1.2" } : plugin);
-          return JSON.stringify({ ok: true, status: "done", name });
-        },
-        PluginDoctor: async (name: string) => {
-          doctorCalls += 1;
-          return { ...(plugins.find((plugin) => plugin.name === name) ?? plugins[0]), warnings: ["manifest exports no MCP auth metadata"] };
-        },
-        RemovePlugin: async (name: string) => {
-          removeCalls += 1;
-          plugins = plugins.filter((plugin) => plugin.name !== name);
-        },
-        PickPluginFolder: async () => {
-          pickFolderCalls += 1;
-          return "/tmp/superpowers-plugin";
-        },
-      } as Partial<AppBindings> as AppBindings,
-    },
-  };
-
-  await act(async () => {
-    root.render(React.createElement(LocaleProvider, null, React.createElement(PluginsSettingsPage)));
-    await flush();
-  });
-  await waitFor("superpowers plugin row", () => Boolean(document.querySelector(".cap-row__name")?.textContent?.includes("superpowers")));
-  ok(Boolean(document.querySelector(".cap-plugin-form-grid .cap-plugin-fields--local")), "local plugin install mode uses the shared form grid");
-  const localOptionTexts = Array.from(document.querySelectorAll(".cap-plugin-installer__options > .cap-plugin-option-block"))
-    .map((option) => option.textContent ?? "");
-  ok(localOptionTexts[0]?.includes("Overwrite same-name plugin"), "local install mode shows overwrite before link mode");
-  ok(localOptionTexts[1]?.includes("Developer mode: link source folder"), "local install mode shows link mode after overwrite");
-
-  const chooseFolder = findButton("Choose plugin folder");
-  if (!chooseFolder) throw new Error("missing plugin folder picker button");
-  await act(async () => {
-    chooseFolder.click();
-    await flush();
-  });
-  await waitFor("picked plugin folder source", () => document.body.textContent?.includes("/tmp/superpowers-plugin") ?? false);
-  ok(pickFolderCalls === 1, "clicking Choose folder invokes the plugin folder picker once");
-
-  const gitMode = findButton("Git repository");
-  if (!gitMode) throw new Error("missing Git repository install mode");
-  await act(async () => {
-    gitMode.click();
-    await flush();
-  });
-  ok(Boolean(document.querySelector(".cap-plugin-form-grid .cap-plugin-fields--git")), "Git plugin install mode uses the shared form grid");
-  const sourceInput = document.querySelector<HTMLInputElement>('input[aria-label="Git repository URL"]');
-  if (!sourceInput) throw new Error("missing plugin git source input");
-  await act(async () => {
-    setInputValue(sourceInput, "git:github.com/obra/superpowers");
-    await flush();
-  });
-  await waitFor("plugin preview enabled", () => findButton("Preview")?.disabled === false);
-
-  const preview = findButton("Preview");
-  if (!preview) throw new Error("missing plugin preview button");
-  await act(async () => {
-    preview.click();
-    await flush();
-  });
-  await waitFor("plugin install plan", () => document.body.textContent?.includes("install_plugin_package") ?? false);
-  ok(planCalls === 1, "clicking Preview invokes plugin install planning once");
-  ok(plannedSources[0] === "git:github.com/obra/superpowers", "plugin preview receives the entered Git source");
-  ok(document.body.textContent?.includes("Partially compatible") ?? false, "preview renders compatibility status");
-  ok(document.body.textContent?.includes("Mapped: skills, agents") ?? false, "preview renders mapped capabilities");
-  ok(document.body.textContent?.includes("hook: unsupported event") ?? false, "preview renders skipped capability reasons");
-
-  const install = findButton("Install plugin");
-  if (!install) throw new Error("missing plugin install button");
-  await act(async () => {
-    install.click();
-    await flush();
-  });
-  await waitFor("plugin install result", () => installCalls === 1 && plugins[0]?.version === "0.1.1");
-  ok(installedSources[0] === "git:github.com/obra/superpowers", "plugin install receives the entered Git source");
-
-  const disclosure = document.querySelector<HTMLButtonElement>(".cap-plugin-entry .cap-disclosure");
-  if (!disclosure) throw new Error("missing plugin disclosure");
-  await act(async () => {
-    disclosure.click();
-    await flush();
-  });
-  await waitFor("plugin update action", () => Boolean(findButton("Update")));
-  ok(document.body.textContent?.includes("How to use") ?? false, "expanded plugin details explain how to use the plugin");
-  ok(document.body.textContent?.includes("/superpowers:plan") ?? false, "expanded plugin details list qualified skill invocations");
-  ok(document.body.textContent?.includes("/superpowers:plan") ?? false, "plugin details show the canonical qualified invocation");
-  ok(document.body.textContent?.includes("qualified name is occupied by a user or project command") ?? false, "occupied canonical command explains the winning source");
-  ok(document.body.textContent?.includes("SessionStart") ?? false, "expanded plugin details list exported hooks");
-  ok(document.body.textContent?.includes("Fully compatible") ?? false, "plugin details show structured compatibility");
-  ok(document.body.textContent?.includes("/superpowers:reviewer") ?? false, "plugin details list imported agents");
-  ok(document.body.textContent?.includes("Context Search") ?? false, "plugin details retain MCP display names");
-  ok(document.body.textContent?.includes("on demand") ?? false, "imported MCP servers are labeled on demand");
-  ok(document.body.textContent?.includes("context") ?? false, "expanded plugin details list exported MCP servers");
-
-  const update = findButton("Update");
-  if (!update) throw new Error("missing plugin update button");
-  await act(async () => {
-    update.click();
-    await flush();
-  });
-  await waitFor("plugin update call", () => updateCalls === 1 && plugins[0]?.version === "0.1.2");
-
-  const doctor = findButton("Doctor");
-  if (!doctor) throw new Error("missing plugin doctor button");
-  await act(async () => {
-    doctor.click();
-    await flush();
-  });
-  await waitFor("plugin diagnostic warning", () => document.body.textContent?.includes("manifest exports no MCP auth metadata") ?? false);
-  ok(doctorCalls === 1, "clicking Doctor invokes plugin diagnostics once");
-
-  const toggle = document.querySelector<HTMLInputElement>(".cap-plugin-entry .cap-switch input");
-  if (!toggle) throw new Error("missing plugin enable toggle");
-  await act(async () => {
-    toggle.click();
-    await flush();
-  });
-  await waitFor("plugin disabled", () => toggleCalls === 1 && plugins[0]?.enabled === false);
-
-  const remove = findButton("Remove plugin");
-  if (!remove) throw new Error("missing plugin remove button");
-  await act(async () => {
-    remove.click();
-    await flush();
-  });
-  const confirmRemove = findButton("Confirm remove");
-  if (!confirmRemove) throw new Error("missing plugin confirm remove button");
-  await act(async () => {
-    confirmRemove.click();
-    await flush();
-  });
-  await waitFor("plugin removed", () => removeCalls === 1 && plugins.length === 0);
 
   await act(async () => {
     root.unmount();
