+++ desktop/frontend/src/__tests__/theme-pack.test.ts
@@ -544,7 +544,8 @@ ok(themePackKind({ builtin: false }) === "user", "legacy builtin=false falls bac
 ok(overviewSource.includes("appearance-overview"), "appearance overview present");
 ok(overviewSource.includes("settings.themeGallery.browse"), "overview has browse themes");
 ok(overviewSource.includes("settings.themeGallery.disable") || overviewSource.includes("handleDisable"), "overview can disable pack");
-ok(settingsSource.includes('tab !== "appearance"'), "appearance renders a single page header");
+const settingsPageShell = settingsSource.slice(settingsSource.indexOf("function SettingsPageShell"), settingsSource.indexOf("export function settingsPageLayout"));
+ok(settingsPageShell.includes("aria-label={settingsTabPageTitle(tab, t)}") && !settingsPageShell.includes("settings-page__header"), "settings pages retain accessible names without a duplicate visual header");
 ok(overviewSource.includes("initialCreateBaseStyle"), "base-style copy opens a prefilled theme editor");
 ok(overviewSource.includes('role="radiogroup"') && overviewSource.includes("aria-checked"), "overview segmented controls expose selection semantics");
 ok(overviewSource.includes("appearance-overview__segmented--theme"), "theme-mode control uses compact settings width");
